% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-utils.R
\name{getAvailableGenes}
\alias{getAvailableGenes}
\title{Get available genes}
\usage{
getAvailableGenes()
}
\value{
A vector of available genes.
}
\description{
This function returns the available genes.
}
\examples{
getAvailableGenes()

}
