% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{initialize}
\alias{initialize}
\title{Initialize function}
\usage{
initialize()
}
\value{
Null; variables are saved to the global environment and accessed by
              the temporary id provided by the API.
}
\description{
This function loads a Seurat object and extracts the expression matrix.
}
\examples{
initialize()
}
