% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{loadSeuratObject}
\alias{loadSeuratObject}
\title{Load a Seurat object}
\usage{
loadSeuratObject()
}
\value{
A Seurat object loaded from "/home/opencpu/data/acp_scn_annotated.rds".
}
\description{
This function loads a Seurat object from a specified file.
}
\examples{
loadSeuratObject()

}
