% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimReduction.R
\name{plotDimReduction}
\alias{plotDimReduction}
\title{Plot Dimension Reduction}
\usage{
plotDimReduction(reduction = "umap", color_by = "celltypes")
}
\arguments{
\item{reduction}{A character string specifying the type of dimension reduction
to plot. Default is 'umap'.}

\item{color_by}{A character string specifying the variable in the metadata
to color cells by. Default is 'celltypes'.}
}
\value{
A data frame with columns 'Dim1', 'Dim2', 'category', and 'color'.
 'Dim1' and 'Dim2' correspond to the two dimensions of the UMAP plot.
  'category' corresponds to the cell type or other category specified
  by 'color_by'. 'color' is determined by the .buildColorRamp function.
}
\description{
This function creates a data frame for plotting dimension reduction methods
such as UMAP or t-SNE.
}
\examples{
plotDimReduction(reduction = 'umap', color_by = 'celltypes')

}
